+++ crates/sui-graphql-rpc/src/context_data/package_cache.rs
@@ -7,7 +7,7 @@ use async_trait::async_trait;
 use diesel::{ExpressionMethods, OptionalExtension, QueryDsl, RunQueryDsl};
 use move_core_types::account_address::AccountAddress;
 use sui_indexer::errors::IndexerError;
-use sui_indexer::{indexer_reader::IndexerReader, schema_v2::objects};
+use sui_indexer::{indexer_reader::IndexerReader, schema::objects};
 use sui_package_resolver::{
     error::Error as PackageResolverError, Package, PackageStore, PackageStoreWithLruCache, Result,
 };
