+++ crates/sui-graphql-rpc/src/data/pg.rs
@@ -189,7 +189,7 @@ mod tests {
     use sui_framework::BuiltInFramework;
     use sui_indexer::{
         get_pg_pool_connection, models_v2::objects::StoredObject, new_pg_connection_pool,
-        schema_v2::objects, types_v2::IndexedObject, utils::reset_database,
+        schema::objects, types_v2::IndexedObject, utils::reset_database,
     };
 
     #[test]
