+++ crates/sui-graphql-rpc/src/types/checkpoint.rs
@@ -24,7 +24,7 @@ use fastcrypto::encoding::{Base58, Encoding};
 use serde::{Deserialize, Serialize};
 use sui_indexer::{
     models_v2::checkpoints::StoredCheckpoint,
-    schema_v2::{checkpoints, objects_snapshot},
+    schema::{checkpoints, objects_snapshot},
 };
 use sui_types::messages_checkpoint::CheckpointDigest;
