+++ crates/sui-graphql-rpc/src/types/protocol_config.rs
@@ -3,7 +3,7 @@
 
 use async_graphql::*;
 use diesel::{ExpressionMethods, QueryDsl};
-use sui_indexer::schema_v2::{checkpoints, epochs};
+use sui_indexer::schema::{checkpoints, epochs};
 use sui_protocol_config::{ProtocolConfig as NativeProtocolConfig, ProtocolVersion};
 
 use crate::{
